Fix incorrect year on January 1.
Since Pepco's dates are missing the year, the first time we collect data on January 1, we get data from December 31.  We think Pepco is saying Dec 31 $(currentYear) but they are actually saying December 31 $(lastYear).
Therefore, if Pepco reports a date of December 31, and it is January 1, we will assume Pepco meant the last year and will set the date/time accordingly.